Relocate sync action
